Fix error with map height
