--- stops/apps/frontend/src/components/Stop/StopMap/Mapper/styles.module.css
@@ -1,19 +1,0 @@
-.container_specific {
-    width: 100%;
-    height: 400px;
-}
-
-.session_specific {
-    width: 100%;
-    height: 400px;
-}
-
-.container_generic {
-    width: 100%;
-    height: 90vh;
-}
-
-.session_generic {
-    width: 100%;
-    height: 90vh;
-}
