+++ stops/apps/frontend/src/components/Stop/StopMap/styles.module.css
@@ -1,3 +1,20 @@
-.container {
+.container_specific {
     width: 100%;
+    height: 400px;
+    min-height: 400px;
+}
+
+.session_specific {
+    width: 100%;
+    height: 400px;
+}
+
+.container_generic {
+    width: 100%;
+    height: 90vh;
+}
+
+.session_generic {
+    width: 100%;
+    height: 90vh;
 }
